refactor: move dummy data to a dedicated directory
